- remove 2 unused sliding menu items - fix mapping json object from API getNilai - fix passing parameter for getClassAndAspect & getNilai

+++ src/com/cox/work/sis/ursula/model/json/NilaiDetilNonRubrik.java
@@ -5,7 +5,7 @@ public class NilaiDetilNonRubrik {
 	public boolean IsRemidi;
 	public JenisNilai JenisNilai;
 	public String Keterangan;
-	public int NilaiAngka;
+	public float NilaiAngka;
 	public int NilaiKe;
 	public float Ratarata;
 	public boolean Status;
